trxcon: trxcon_fsm: permit direct DCH/PDCH transitions
It may happen that the MS needs to go directly from a PDCH to a TCH
or SDCCH, e.g. in case of a Mobile Terminated call.  I don't know
if the opposite transition is needed on practice, but let's simply
allow direct transitions for both TRXCON_ST_{DEDICATED,PACKET_DATA}.

Change-Id: I1a854e4683f102c40f1c174a291b6dc638f49b5c
Related: OS#5599