trxcon: move FBSB timeout calculation to the trxcon_fsm
In the upcoming patches I am adding a possibility to enlarge the FBSB
timeout by providing API for that.  This is needed for SDR based PHYs,
because for them it takes longer to tune and so on.  The L1CTL codec
is not the right place for applying PHY specific quirks, so let's
move the TDMA FNs -> ms conversion to the FSM logic.

Change-Id: I685f48cfed000997b0d7c16073c6387bc05d2bbe
Related: OS#5599