layer23: modem: Unregister registered callbacks upon app exit
It's just a good practice to delete all resources allocated during startup.
The main aim here is to keep resemblance to what the mobile app is doing,
so that they can slowly be merged and some functionalities from the mobile app can be
added to the modem app, like shutting down the MS without killing the process eventually.

Change-Id: I5a641fa3dadb6ea7346b25a20215896ab32eb805