virt_phy: fix bogous TDMA Fn check in l1ctl_rx_gprs_ul_block_req()
sched_fn_ul() does not support RSL_CHAN_OSMO_PDCH, so it would always
return the current time, which in most cases is not the correct time
for scheduling a block.  Actually, we don't really need this function
because the Tx Fn is provided to us by the upper layers - just use it.

Change-Id: Ieb2c819db1ef9377680866989cbb754fac5e45bd
Related: OS#5500