trxcon: do not advance Uplink TDMA Fn by default
The idea behind advancing Uplink TDMA Fn is to give the transceiver,
which is usually a separate process, some additional time to receive
and prepare Uplink bursts for transmission.  This comes at a price
of having an additional delay between Uplink and Downlink.

Given that trxcon, as a standalone application, is primarily used in
conjunction with fake_trx.py for running ttcn3-bts-test against
osmo-bts-trx, there is no reason to advance the Uplink TDMA Fn.

Change-Id: I838b1ebc54e4c5d116f8af2155d97215a6133ba4
Related: OS#5500