layer23: fix handling of logging category mask (-d option)
In change 67943df4 I broke handling of the logging category mask in
the mobile app.  Adding this option results in a segfault:

  ERROR: osmo_log_info == NULL! You must call log_init() before
         using logging in log_parse_category_mask()!
  Assert failed osmo_log_info src/libosmocore/src/core/logging.c:329

As can be seen, the problem is that we are calling
log_parse_category_mask() before initializing the logging.

As possible solution, I could rearrange the code to parse command
line options after calling osmo_init_logging2().  This would fix
the segfault, but would not fully solve the problem.

If we call log_parse_category_mask() before parsing the config file,
then logging configuration in the config file overwrites the logging
configuration specified via the command line.  But we want the
opposite: the command line setting should overwrite the config file
parameters.  This is handy because there is no need to edit the
config file if you quickly need to test something.

So let's call log_parse_category_mask() after parsing the config file.

Change-Id: I1b2b7804bf99b71f96e9197f7824cfd20431e8a1
Fixes: 67943df4 "layer23: fix parsing of command line options"