firmware/layer1: mute UL/DL vocodec if it's not needed
The upper layers usually request either of the two configurations:

* (AUDIO_TX_MICROPHONE | AUDIO_RX_SPEAKER) - in this configuration
  the phone (PHY) is both the origin and the destination of the TCH
  frames.  DL frames are played via the built-in speaker; UL frames
  recorded using the built-in microphone.

* (AUDIO_TX_TRAFFIC_REQ | AUDIO_RX_TRAFFIC_IND) - in this case
  the upper layers (host side) become the origin and the destination
  of the TCH frames.  The built-in speaker and microphone are
  expected to be disabled.

However, when using the second configuration, one can still hear
DL TCH frames being played by the built-in speaker.  The built-in
microphone does not seem to be causing any issues, but still we
definitely don't want the vocoder to interfere with the host.

Change-Id: I390db1889f079dea8112794c3e039a9136b897df
Related: OS#4396