trxcon/l1sched: refactor prim management in tx_tch[fh]_fn()
The code path below the switch statement in tx_tch[fh]_fn() is no
longer common since we added CSD specific channel coding.  This is
why we had to jump over it in several case statements.

This patch significantly reduces the number of goto statements
in these two functions and makes them easier to read/follow at
the price of code duplication, which is tolerable.

Change-Id: I5292abf6fcd308c9f7f12c7145d004103c9c7675