trxcon/l1sched: fix FACCH/H regression in rx_tchh_fn()
In c15084a5 overlooked that in order to ensure alignment to the first
FACCH/H block in rx_tchh_fn() we actually need to check if DL FACCH/H
can start (not end!) at the current TDMA Fn.  This means we cannot use
the same mapping as we do below in that function;  we need another one.

This patch fixes multiple FACCH/H regressions in ttcn3-bts-test.

Change-Id: Ia4b737cf11d4d9ce9847cabb77189e9cbcbb8840
Fixes: c15084a5 ("trxcon/l1sched: replace old API with sched_tchh_ul_facch_map[]")